Update Preset Parser and Presets.json for : Using Harbor BlockedArea Buildings (#323) Using Spanish Language Using decimals for Influence Radius (GU12 #346) Fixed Missing Influence Radius for  - Iron Tower (#351)  - Harbor Master Office  - Repair Cranes Fixed Enbesa Salt Works icon Update treeLocalization.json (1.0.0.5) Update Presets.json (version 4.0)
Merge this PR if teh PR #354 is merged!!

+++ PresetParser/Models/ExtraPreset.cs
@@ -9,7 +9,7 @@
         public string IconFileName { get; set; }
         public int BuildBlockerX { get; set; }
         public int BuildBlockerZ { get; set; }
-        public int InfluenceRadius { get; set; }
+        public double InfluenceRadius { get; set; }
         public int InfluenceRange { get; set; }
         public string Template { get; set; }
         public string LocaEng { get; set; }
